+++ source/Classic/Scenes/RestaurantScene.cs
@@ -61,6 +61,13 @@ namespace Burntime.Classic.Scenes
             font = new GuiFont(BurntimeClassic.FontName, new PixelColor(212, 212, 212));
         }
 
+        public override void OnResizeScreen()
+        {
+            base.OnResizeScreen();
+
+            Position = (app.Engine.Resolution.Game - new Vector2(320, 200)) / 2;
+        }
+
         protected override void OnActivateScene(object parameter)
         {
             BurntimeClassic classic = app as BurntimeClassic;