+++ source/Classic/Scenes/StatisticsScene.cs
@@ -38,6 +38,13 @@ namespace Burntime.Classic.Scenes
             CaptureAllMouseClicks = true;
         }
 
+        public override void OnResizeScreen()
+        {
+            base.OnResizeScreen();
+
+            Position = (app.Engine.Resolution.Game - new Vector2(320, 200)) / 2;
+        }
+
         public override void OnRender(RenderTarget Target)
         {