+++ source/Classic/Scenes/TraderScene.cs
@@ -76,6 +76,13 @@ namespace Burntime.Classic.Scenes
             Windows += temporarySpace;
         }
 
+        public override void OnResizeScreen()
+        {
+            base.OnResizeScreen();
+
+            Position = (app.Engine.Resolution.Game - new Vector2(320, 200)) / 2;
+        }
+
         public override void OnRender(RenderTarget Target)
         {
             base.OnRender(Target);